requirements.txt
requests
allure-pytest
jsonpath
pymysql
pytest
PyYaml